Move GetServerVersion with other logic related to fetching cluster info
